TaterSetListManager for GrandMA3 V2.2.5
Readme! : Your songmacros need to have the same exact name as the sequences for this to work (Macro Song2 will reference Sequence Song2, etc)

All of the "last cues" need to have the same cue number, by default it is cue 500. You can edit this on the user interface! The current set lastcue will be in paranthesis on the set last cue button.

This plugin works by taking a range of macros, it reads the name and order of the macros, and it will update the last cue of each song to update to a new order. To reorder the setlist, simply change the order of the macros. The first macro you select will be the first song, and it counts up sequentially until the last macro that you set.

The plugin will not update the lastcue until you press update, it will update the setlist on the user interface so that you can double check the order is correct.
Please send any bugs to Tater@LXTater.com
Updated 4/6/2025
MA3 V 2.2.5.0