CSV to 3D Plugin Creator for MA3
https://lxtater.com/csvtool/

I was annoyed with the inability to upload a .mvr into MA3 3D that already has fixtures, and only merge 3D position information. This allows you to export a .csv from your pre-vis software, and it will create a plugin that will allow you to set your current fixtures to the 3D positions in the CSV. Please note: The plugin goes off of Fixture IDs. This tool reads either channel or fixture ID in the CSV as the Fixture ID. This won't work properly if you have both a channel and fixture ID set that are different than the MA3 Patch. As always, SAVE before you run this plugin!

I will be including the ability to directly upload .mvr soon!